// vscode/src/generated/diagram-product/DiagramEdge.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Provenance } from "./Provenance";

export type DiagramEdge = { reference: number, source: number, target: number, origin: number, kind: string, provenance: Provenance, navigation: number | null, };