// vscode/src/generated/diagram-product/DiagramNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Compartment } from "./Compartment";
import type { NodeTyping } from "./NodeTyping";
import type { NotationRole } from "./NotationRole";

export type DiagramNode = { reference: number, metaclass: string, notationRole: NotationRole, name: string | null, typing: NodeTyping, owner: number | null, source: number, compartments: Array<Compartment>, direction: string | null, conjugated: boolean, };