// vscode/src/generated/diagram-product/SemanticReference.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SourceDomain } from "./SourceDomain";

export type SemanticReference = { "kind": "qualified-name", document: number, qualifiedName: string, } | { "kind": "tooling-element-id", elementId: string, sourceDomain: SourceDomain, } | { "kind": "source-anchor", source: number, ownerQualifiedName: string | null, metaclass: string, sourceDomain: SourceDomain, } | { "kind": "relationship", source: number, relationshipKind: string, ordinal: number, };